+++ src/components/camel-app-resources/podStatus.ts
@@ -1,0 +1,80 @@
+import * as _ from 'lodash';
+
+const isContainerFailedFilter = (containerStatus) => {
+  return containerStatus.state.terminated && containerStatus.state.terminated.exitCode !== 0;
+};
+
+export const isContainerLoopingFilter = (containerStatus) => {
+  return (
+    containerStatus.state.waiting && containerStatus.state.waiting.reason === 'CrashLoopBackOff'
+  );
+};
+
+const numContainersReadyFilter = (pod) => {
+  const {
+    status: { containerStatuses },
+  } = pod;
+  let numReady = 0;
+  _.forEach(containerStatuses, (status) => {
+    if (status.ready) {
+      numReady++;
+    }
+  });
+  return numReady;
+};
+
+const isReady = (pod) => {
+  const {
+    spec: { containers },
+  } = pod;
+  const numReady = numContainersReadyFilter(pod);
+  const total = _.size(containers);
+
+  return numReady === total;
+};
+
+const podWarnings = (pod) => {
+  const {
+    status: { phase, containerStatuses },
+  } = pod;
+  if (phase === 'Running' && containerStatuses) {
+    return _.map(containerStatuses, (containerStatus) => {
+      if (!containerStatus.state) {
+        return null;
+      }
+
+      if (isContainerFailedFilter(containerStatus)) {
+        if (_.has(pod, ['metadata', 'deletionTimestamp'])) {
+          return 'Failed';
+        }
+        return 'Warning';
+      }
+      if (isContainerLoopingFilter(containerStatus)) {
+        return 'CrashLoopBackOff';
+      }
+      return null;
+    }).filter((x) => x);
+  }
+  return null;
+};
+
+export const getPodStatus = (pod) => {
+  if (_.has(pod, ['metadata', 'deletionTimestamp'])) {
+    return 'Terminating';
+  }
+  const warnings = podWarnings(pod);
+  if (warnings !== null && warnings.length) {
+    if (warnings.includes('CrashLoopBackOff')) {
+      return 'CrashLoopBackOff';
+    }
+    if (warnings.includes('Failed')) {
+      return 'Failed';
+    }
+    return 'Warning';
+  }
+  const phase = _.get(pod, ['status', 'phase'], 'Unknown');
+  if (phase === 'Running' && !isReady(pod)) {
+    return 'NotReady';
+  }
+  return phase;
+};
