+++ src/const.ts
@@ -2,21 +2,6 @@ import { K8sGroupVersionKind } from '@openshift-console/dynamic-plugin-sdk';
 
 export const HAWTIO_CONSOLE_PLUGIN_NAME = 'hawtio-online-console-plugin';
 
-export const METADATA_LABEL_SELECTOR_CAMEL_APP_KEY = 'camel/integration-runtime';
-export const METADATA_LABEL_SELECTOR_CAMEL_APP_VALUE = 'camel';
-
-export const METADATA_ANNOTATION_APP_VERSION = 'app.kubernetes.io/version';
-
-export const METADATA_ANNOTATION_CAMEL_VERSION = 'camel/camel-core-version';
-
-export const METADATA_ANNOTATION_CAMEL_QUARKUS_PLATFORM_VERSION = 'camel/quarkus-platform';
-export const METADATA_ANNOTATION_CAMEL_CEQ_VERSION = 'camel/camel-quarkus';
-
-export const METADATA_ANNOTATION_QUARKUS_BUILD_TIMESTAMP = 'app.quarkus.io/build-timestamp';
-
-export const METADATA_ANNOTATION_CAMEL_SPRINGBOOT_VERSION = 'camel/spring-boot-version';
-export const METADATA_ANNOTATION_CAMEL_CSB_VERSION = 'camel/camel-spring-boot-version';
-
 export const camelAppGVK: K8sGroupVersionKind = {
   group: 'camel.apache.org',
   version: 'v1alpha1',
@@ -83,3 +68,4 @@ export const consolePluginGVK: K8sGroupVersionKind = {
 };
 
 export const ALL_NAMESPACES_KEY = '#ALL_NS#';
+export const LAST_LANGUAGE_LOCAL_STORAGE_KEY = 'bridge/last-language';
